requirements-functions.txt
# Azure Functions requirements for msgraph-delta-query with Azure Blob Storage
# This file is for Azure Functions deployment with all dependencies

# Core Microsoft Graph delta query dependencies
aiohttp>=3.8.0
azure-identity>=1.12.0

# Azure Blob Storage for delta link persistence
azure-storage-blob>=12.14.0

# Optional: Additional Azure dependencies that might be useful
# azure-functions>=1.11.0  # Usually provided by Azure Functions runtime
# azure-keyvault-secrets>=4.7.0  # If using Key Vault for secrets

# Optional: Environment/configuration management
python-dotenv>=1.0.0  # For local development with .env files

# Optional: Enhanced logging and monitoring  
# azure-monitor-opentelemetry>=1.0.0  # For Application Insights integration
